lib/ploop.c: fix 32-bit compiler warning
ploop.c: In function ‘ploop_resize_image’:
ploop.c:1499:4: error: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 7 has type ‘off_t’ [-Werror=format]
cc1: all warnings being treated as errors

+++ lib/ploop.c
@@ -1499,7 +1499,7 @@ int ploop_resize_image(struct ploop_disk_images_data *di, struct ploop_resize_pa
 			ploop_err(0, "Unable to change image size to %llu "
 					"sectors, minimal size is %lu",
 					param->size,
-					(dev_size - available_balloon_size));
+					(long unsigned)(dev_size - available_balloon_size));
 			ret = SYSEXIT_PARAM;
 			goto err;
 		}
